requirements.txt
python-telegram-bot>=20.0
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
